Fix on all Tests according to the review

+++ src/PaymentsClient.Domain/NagApiSettings.cs
@@ -5,5 +5,5 @@ public record NagApiSettings
     public Uri? BaseUri { get; set; }
     public string? ClientId { get; set; }    
     public string? ClientSecret { get; set; }
-    public int? TimeOutInSeconds { get; set; }
+    public double? TimeOutInSeconds { get; set; }
 }
